refactor(web): move cn utility inside shadcn for modularization
