+++ web/components/shadcn/button/Button.vue
@@ -2,7 +2,7 @@
 import type { HTMLAttributes } from 'vue'
 import { Primitive, type PrimitiveProps } from 'radix-vue'
 import { type ButtonVariants, buttonVariants } from '.'
-import { cn } from '@/helpers/utils'
+import { cn } from '~/components/shadcn/utils'
 
 interface Props extends PrimitiveProps {
   variant?: ButtonVariants['variant']
