+++ web/components/shadcn/scroll-area/ScrollArea.vue
@@ -7,7 +7,7 @@ import {
   ScrollAreaViewport,
 } from 'radix-vue'
 import ScrollBar from './ScrollBar.vue'
-import { cn } from '@/helpers/utils'
+import { cn } from '~/components/shadcn/utils'
 
 const props = defineProps<ScrollAreaRootProps & { class?: HTMLAttributes['class'] }>()
